Begin changelog entry for the next version
This contains both current and expected changes. More to come, probably.

Barring any major problems in 2109-9 necessitating a sooner release, this will probably all be built once Linux 5.14 and Plasma 5.22 enter Testing. I don't plan to do rebuilds on a consistent schedule, I'll just do it when we either need to support new hardware, push meaningful changes that new installations should have ASAP, and/or it's been long enough that the size of the post-install upgrade is annoying.

Linux 5.14 is currently in Experimental. Plasma 5.22 is in VCS, awaiting upload. Might come out mid-October?

The version scheme is slightly changing, from YEARMONTH-BUILDNUMBER to simply YEAR-MONTH now that we're probably not going to be making multiple releases in a month, making the next version Khromian 21-10 most likely. Though, if a hotfix is necessary, a build number will still be affixed to distinguish between patched and unpatched images.