More ZRam tweaking...
After testing on a VM with only 1GB of RAM, I'm comfortable using the entire physical memory, and I think I misunderstood how it worked before. It was also a surprisingly good experience, even on a terrible CPU and with only 1GB of RAM, which is a huge accomplishment I think.

I've also tested NoHang tweaks. I think it needs to be more aggressive. I got the configuration into a good spot where it would push the limits of ZRAM and physical memory until a freeze happens for multiple seconds, where it would trigger. I'm happy with that for desktop use.