KIO Fuse included + Chromium on all image types
*Not* including KIO Fuse was an oversight I missed until now. Glad to have that in here.

The browser issue is a tricky one. I think it's ultimately more confusing to use a different browser on our minimal images though. Chromium still has better KDE integration, and better out-of-the-box support for hardware video acceleration, which is probably particularly important on the types of systems people might use the minimal image on.