ZRAM tweaks, and more!!!
I've discovered that in Linux 5.8>, you can set vm.swappiness to 200 to make extremely heavy use of swap space (compressed RAM pages in our case). That sounds apt for us, as I've been struggling to find a scenario where ZRAM has any negative impact on performance that isn't completely outweighed by the reduction in the size of data that has to be read and written, i.e. even if the (de)compression speed is 3x slower, it would still level out if it's writing 1/3rd of the data, and often that ratio is in our favor. I might reduce this if it seems to be overkill, but it has made a terrible VM extremely snappy right now, so I'm going to start out overconfident with it. 

Additionally, using a wider variety of tests has shown that creating a larger ZRAM device than your physical RAM doesn't actually appear to degrade performance when nearing the limit, unless the system is already extremely choked for CPU. Though, in that case, it should hopefully be handled by our aggressive NoHang configuration. 200% of the physical RAM size appears to be a generous yet safe choice.

On an unrelated note, we'll have to backpedal on including links to the changelog and known issues in the installer, as it currently attempts to open Chromium as root to do this, which is obviously insecure, and something that causes Chromium to fail on startup anyhow.

We can also lower the storage requirement even further to only 6.5GiB, as testing shows that extraction only peaks at about 5.7GiB of usage still. Keep in mind that we also need to reserve 500M for the ESP, which is why it isn't a round 6GiB.

Finally, I'm pretty confident in pushing a 2109-10 build, as the last in the 2109 series. It'll take a while to generate and test all the extra images though, so I don't know when it'll roll out exactly. This changelog should be basically finalized though. If there's any possible reversion at this point, it would be using the default PipeWire session manager instead of WirePlumber, as WP may or may not have issues correctly detecting what the default audio profile for a device should be. Needs further testing.