Additions and removals...
Don't know why we weren't already bringing in plasma-firewall and plasma-runner-launcher as they're very handy. On the other hand, plasma-wayland-protocols is a confusing one. It's not pulled in by plasma-workspace-wayland but the description of the package makes it sound important. It's so small that I have no qualms about pulling it in, just in case a user wants a Wayland session later and it turns out to be important.

ImageMagick will haunt my pure application list no longer.