Finalize 21-10 changelog, hopefully
Few final changes, the largest one being that NVIDIA drivers are pulled from Sid which might be a bad idea and hopefully won't come back to bite me in the ass. Should be ready to ship though. Good chance I'm releasing a hotfix before the end of the month though.